\name{GLNetExample}
\alias{GLNetExample}
\docType{data}
\title{
	An Example of Gene-LncRNA Network. 
}
\description{
	An adjaency matrix of gene-lncRNA network. An Example of gene-lncRNA Network, which would be used to construct multi-omics composite network.
}
\usage{data("GLNetExample")}

\examples{
data(GLNetExample)
## maybe str(GLNetExample) ; 
}
\keyword{datasets}
