\name{GNetExample}
\alias{GNetExample}
\docType{data}
\title{
	An Example of Gene Network. 
}
\description{
	An adjaency matrix of gene network. An Example of gene Network, which would be used to construct multi-omics composite network.
}
\usage{data("GNetExample")}

\examples{
data(GNetExample)
## maybe str(GNetExample) ; 
}
\keyword{datasets}
