\docType{package}
\name{LncPriNet-package}
\alias{LncPriNet-package}
\title{Description of LncPriNet Package}
\description{A R platform to prioritize candidate disease lncRNAs by LncPriNet method which extend random walk analysis on multi-omics network.}
\section{Introduction}{
	The main goal of LncPriNet is to prioritize candidate disease lncRNAs by LncPriNet method. 

}

\author{ Qianlan Yao <yaoqianlan@yahoo.com> }
