\name{LNetExample}
\alias{LNetExample}
\docType{data}
\title{
	An Example of LncRNA Network. 
}
\description{
	An adjaency matrix of lncRNA network. An Example of lncRNA Network, which would be used to construct multi-omics composite network.
}
\usage{data("LNetExample")}

\examples{
data(LNetExample)
## maybe str(LNetExample) ;
}
\keyword{datasets}
