\name{PGNetExample}
\alias{PGNetExample}
\docType{data}
\title{
	An Example of Phenotype-gene Network. 
}
\description{
	An adjaency matrix of phenotype-gene network. An Example of Phenotype-gene Network, which would be used to construct multi-omics composite network.
}
\usage{data("PGNetExample")}

\examples{
data(PGNetExample)
## maybe str(PGNetExample) ;
}
\keyword{datasets}
