\name{PLNetExample}
\alias{PLNetExample}
\docType{data}
\title{
	An Example of Phenotype-lncRNA Network. 
}
\description{
	An adjaency matrix of phenotype-lncRNA network. An Example of Phenotype-lncRNA  Network, which would be used to construct multi-omics composite network.
}
\usage{data("PLNetExample")}

\examples{
data(PLNetExample)
## maybe str(PLNetExample) ;
}
\keyword{datasets}
