\name{PNetExample}
\alias{PNetExample}
\docType{data}
\title{
	An Example of Phenotype Network. 
}
\description{
	An adjaency matrix of phenotype network. An Example of Phenotype Network, which would be used to construct multi-omics composite network.
}
\usage{data("PNetExample")}

\examples{
data(PNetExample)
## maybe str(PNetExample) ;
}
\keyword{datasets}
